+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - resultPrimary dataProfile.
 - resultBroad dataProfile.
 - added organization summary.
-- added period of record data profile.
+- added period of record data profile with CSV output.
 
 ### Changed
 - Station summary service returns only sites that have summary information.